Bug 2228: crash closing, only Linux, gcc, opt, wxWidgets 3.0...
... Not the most satisfactory fix, but in fact no dangling pointers will
happen with the code we have now, because the relevant windows are all destroyed
only at the end of AudacityProject's lifetime, including the non-modal windows
(macros, history, lyrics, mixer board, plot spectrum, and contrast) which
are hidden and shown again, not destroyed and recreated, when dismissed and
reopened.

To do:  figure out how to make wxWeakRef work on that combination without
crashing.