Implement Multi View checkbox menu item in wave track TCP...
... When it's off (default), Spectrum and Waveform buttons behave as before,
and dragging of the sub-view separator is disabled.

When it transitions to off, and the view is split, then the top sub-view takes
up the whole view.

When it transitions on, nothing visible happens.

When it is on, and you choose Spectrum or Waveform, then the corrsponding
sub-view toggles visibility.

When a sub-view is turned on by the menu item, it appears lowest.