Provides to use the system wxWidgets install
The default is to use the system one if found.  If the
WXWIN environment variable is set and points to a valid
wxWidgets tree, that will be used.  Otherwise, it will
search the platform specific locations.

You can use a local library instead with:

   cmake -Duse_system_wxwigets=no ...

In this case it will look for the WXWIN environment variable
and use that or it will download the Audacity specific wxWidgets
source.  In both cases, wxWidgets will be built as part of the
Audacity project.