Fix the Github CMake build
Restore the wxDEBUG_LEVEL defines for the module builds.  It is
currently required (see explanation in CMakeLists.txt)

Add explanantion of why wxWidgets_USE_REL_AND_DBG=NO is needed
for the Github Windows build.

Further restrict the disabling dynamic loading for Windows and Mac
of the Lame and FFmpeg libs.  Some corrections need to be made in
ExportMP3.cpp (and probably some for FFmpeg as well).

And don't build mod-nyq-bench or mod-null by default until they
can be fixed.  (They fail to build.)