Define and use SettingScope and SettingTransaction...
... This utility may find more uses.  It makes it easy to ensure that cancelling
the Preferences dialog restores all preferences as they were when the dialog
was constructed.  This could simplify the previewing logic for Spectrogram
preferences, but this is not yet done.

Also in the image compiler, using it leaves GUIBlendThemes undefined if it was
previously undefined.  This improved behavior will only matter when you build
Audacity for yourself.