[CMAKE] PCH: wxWidgets must be included first
To make working the support of precompiled headers, it is required that wxWidgets headers are included before all others. This is required because they can activate some options before the inclusion of standard libc files. Otherwise, if you include other files before those options are activated, the build process may crash or the application may crash during runtime. I noticed this behavior when building on Cygwin, that requires the presence of macro _GNU_SOURCE to be defined. Without that, the build process will fail with these errors:

src/AudioIO.cpp: In function 'double SystemTime(bool)':
src/AudioIO.cpp:817:21: error: 'CLOCK_MONOTONIC_RAW' was not declared
in this scope; did you mean 'CLOCK_MONOTONIC'?
  817 |       clock_gettime(CLOCK_MONOTONIC_RAW, &now);

src/FileNames.cpp: In function 'FilePath FileNames::PathFromAddr(void*)':
src/FileNames.cpp:428:4: error: 'Dl_info' was not declared in this scope
  428 |    Dl_info info;

For Reference, this is the message sent to the mailing list:

https://sourceforge.net/p/audacity/mailman/audacity-devel/thread/CADt9577Z0oDHB2nahM7%2BboRNEBFyYaX_NE2UV5d9f%2BASTGReTg%40mail.gmail.com/#msg36938976