[WIN32] Correct declaration of import/export modifiers.
The '_declspec' modifier is deprecated since the migration from Win16 to Win32, so by long time.
The correct modifier to be used for modern code is '__declspec', with two underscore characters as prefix.
MSVC can recognize both modifiers and '_declspec' has the same effect of '__declspec', depending on the presence of '/Za' option, which activates/deactivates the language extensions:

https://docs.microsoft.com/it-it/cpp/cpp/declspec?view=vs-2019

Unfortunately, GCC supports only '__declspec', so I recommend to use this one and also MSVC will be able to work indipendently from the language extensions. Since Audacity does not compile on anything less than GCC 4.9, it is not a big problem to use the same thing for _WIN32 in general, not only MSVC, and also Cygwin.