Bug 2334 - Mac: AU effects cannot Export or Import the effect's settings
This removes the ability to import/export multiple presets at
once and add the standard method used by the other plug-ings.

It also adds a workaround for the AU view not updating after
setting a new preset.  I will continue to look for the true
cause of this, but the workaround does correct the display
issue.