Prevent random activations of the Pause button
I don't know how this hasn't been seen before, but Cliff found that
opening Audacity via Commander One would, more often than not, cause
the pause button to enable when you started monitoring.

Seems it was picking up uninitialized memory values, but near as I can
tell, this has been this way for a very long time (like forever).  So,
it's possible that I've not completely identified the culprit.

Even so, this should be initialized anyway and Cliff reports that it
does resolve the problem for him.