Fix crash on track resize
As discussed in #538, there is an issue with the track resize option,
where with audacity >= 2.4.0, compiled using the autotools setup leads
to a hard crash.

The sizeof(wxFrame) in src/menus/ViewMenus.cpp and src/ProjectWindow.cpp
is not aligned, leading to GetVerticalScrollBar in
src/menus/ViewMenus.cpp accessing unallocated memory.

By ordering the include of ProjectWindow.h before the include of
Project.h, this crash is prevented.