Bug 2499 - Macros: applying macro to project focus problem
Problem:
Applying macro to project always sets first track as focus.

The problem was introduced by commit 06cddda, which accidentally declared cleanup2 in the wrong scope.
This meant that ProjectHistory::RollBackState() was always called, which eventually causes TrackPanel::OnUndoReset() to be called, which sets the first track to be the focus.

Fix:
Declare cleanup2 in the correct scope.