Make waveform navigation via mouse pan/zoom simpler
Makes zooming and panning simpler by slowing down the amount of
pan (by 50%) and zoom (by 90%) per event. Additionally, it clamps
zooming out to fit the screen: once the waveform fits on the screen it
will not zoom out any further with the mouse. This prevents the
waveform getting "lost", which is confusing.

The zoom-out command (and keyboard shortcuts) will still zoom out
further if the user wants a more compressed view (eg, to paste audio
at a future location) and mouse zooming will work as expected when
zoomed out.

The speed changes make it easier to control movement and target areas
of the files. On my test machines (Mac, Linux, with trackpads and
external mice) it is still possible to go from absolute minimum to
absolute maximum zoom with a single gesture - but it is far easier to
make precise fine adjustments.

I have tested this commit with short, long, and very long
files/tracks - single track and multitracks, and I feel like they
make "nicer defaults" (especially for beginners or new Audacity
users). It feels more intuitive (to me), and alleviates my personal
pain points - though I have not tested it on Windows where the mouse
handling may be different.