More format fixes on import: beware sequences with differing format...
... It could happen.  When collecting file information to be used in the
second pass, record the correct format for each.

Don't mistakenly use whatever was the format of the last read Sequence for all
files.

This only matters for the avoidance of unnecessary sample format conversions
when finally appending samples to tracks.  It won't affect the space ultimately
occupied in the database.