Other TypeSwitch details don't use a closed-ended type list...
... and don't use C-style pointer casts but instead static_cast.

Track::TypeSwitch hides the pointer-casting from you so you can just supply
several lambdas.  It resembles std::visit perhaps but also allows
open-endedness of the set of subtypes of Track.

And I still don't like it and think most uses of it should be replaced with
new virtual functions of Track, or with AttachedVirtualFunction.  But it's not
in scope of the present project to eliminate all of them.