At last implement the changed behavior of Append() and Set()...
... And construct empty WaveClips assuming dither is not needed, until appending
or setting content determines otherwise.

Results of resampling still dither (happens below WaveTrack level; see
WaveClip.cpp).

This also complete the review of ALL calls to CopySamples everywhere for their
dithering effects.

(Note call to AppendNewBlock from ImportAUP did not dither before, because the
buffer format was same as the format sequence was constructed with; and it will
not happen now either.)