Bug2605: Save-as should not cause original to repoen with warnings...
... Problem was that before abandoning the connection to the original path,
compaction was done only sometimes, according to the criterion of estimated
%20 space savings at least.  But this could cause the project to reopen with
orphan blocks in it, which were detected and reported as an improper save.

The easy solution is to force compaction always in this case.

Direct save does not have this problem because what would otherwise be orphaned
blocks are directly deleted from the database when undo states are destroyed
and sample blocks become unreferenced.  So compacting only sometimes does not
cause orphan blocks on reopening.