Lift some calculation out of OutputEvent() and GetNextEvent()...
... Which will be consumer-thread side, that won't make the decisions about
which parts of the track to play.

They stop using the function MidiLoopOffset(), which will be eliminated.

Instead some information is stored in the iterator, and this may need updating
by the producer even after the consumer has made some use of the iterator.  So
use atomics in the iterator.