Bug 2648: Grid accessibility: location can be incorrect.
The focus location provided to assistive technology by GridAx::GetLocation() does not take into account horizontal or vertical scrolling. So if there is any scrolling in either the Metadata or Labels editor, the focus location provided to assistive technology is not the same as the focus location shown by wxWidgets.