Bug 2640 - Plot Spectrum and Contrast analyzers are not "Repeatable"
Completing the update by John Colket.  The plan always was that
these new 'repeat' entries were optional commands bindable to
keyboard shortcuts by advanced users.

Users visiting the menus can re-select the item they want.
For these menu users it would be confusing/overload to list
these repeat commands in the menu.  Repeat last effect in the Effects
menu is more important.  It is there in the menu to signal to
users that there is the Ctrl+R key binding for it.