Bug 2665 - Docimage script now runs ahead and stalls.
There are now extra blank lines in the results that
need to be dealt with.  Script updated to do so.

This bug 2665 is now fixed.  I think there is a small issue
with the label command not using the selection that
still needs to be addressed too, but it is minor compared
to bug 2665 and can wait.