Add Comments (for Bug 2673 fix)
Bug 2673 was an important multi-threading issue, and so I added
comments about the code.  Also a LogDebug to track cache use.

Bug 2673 was also a crash with no error message reported.
That is something to revisit later and deserved a comment.
It's believed that the underlying issue is fixed now.

The 'ANSWER-MEs' can be answered in 3.0.1.