Fix (workaround) GUI hang on Linux
On second display of LV2 effect using an X11UI, Audacity
would hang due to the glib class not being in the expected
state.

Workaround is to not unload the suil wrapper library. The
unload was already being bypassed on Windows so I'm unsure
why the author didn't do it on the other platforms.