Add C number format option (disabled)
The option to force decimal point in numbers and not
use comma is an experiment to work around bug 2709

It's disabled, because wxWidgets checks for a mismatch
between locale and number format, and disallows it.  To get
it working correctly might require patching wxWidgets.

This change does get the option string into our source code
so that translators can translate it.