Bug 2676 - Export: Incorrect handling over 0 dB
This bug was caused by dither being applied to all float output,
and dither clips to +1/-1 to avoid high values, especially NaNs,
polluting audio downstream.

Dither was applied because of an incorrect fix for Bug 1572.
The 1572 fix assumed SF_FORMAT_PCM_24 was a  (1 << N)
value, rather than from an enumeration.  Hence a check
involving it requires a mask, which it now has.